Rag/requirements.txt
pypdf>=3.17.0
langchain>=0.1.0
langchain-community>=0.0.10
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.2
chromadb>=0.4.18
accelerate>=0.25.0
bitsandbytes>=0.41.0  # For 8-bit quantization
python-dotenv>=1.0.0  # For email configuration
fastapi>=0.109.0  # For REST API
uvicorn>=0.27.0  # ASGI server for FastAPI
pydantic>=2.0.0  # For data validation 